LoadingViewを月ファイルに切り出し
